moving stan/math/prim/scal/fun/inverse_softmax.hpp to arr; moving stan/math/prim/mat/err/check_matching_sizes.hpp to arr
