+++ stan/math/prim/arr.hpp
@@ -11,12 +11,14 @@
 #include <stan/math/prim/arr/meta/VectorBuilderHelper.hpp>
 #include <stan/math/prim/arr/meta/VectorView.hpp>
 
+#include <stan/math/prim/arr/err/check_matching_sizes.hpp>
 #include <stan/math/prim/arr/err/check_nonzero_size.hpp>
 #include <stan/math/prim/arr/err/check_ordered.hpp>
 
 #include <stan/math/prim/arr/fun/dot.hpp>
 #include <stan/math/prim/arr/fun/dot_self.hpp>
 #include <stan/math/prim/arr/fun/fill.hpp>
+#include <stan/math/prim/arr/fun/inverse_softmax.hpp>
 #include <stan/math/prim/arr/fun/log_sum_exp.hpp>
 #include <stan/math/prim/arr/fun/promote_scalar.hpp>
 #include <stan/math/prim/arr/fun/promote_scalar_type.hpp>
