+++ stan/math/prim/arr/err/check_matching_sizes.hpp
@@ -1,5 +1,5 @@
-#ifndef STAN_MATH_PRIM_MAT_ERR_CHECK_MATCHING_SIZES_HPP
-#define STAN_MATH_PRIM_MAT_ERR_CHECK_MATCHING_SIZES_HPP
+#ifndef STAN_MATH_PRIM_ARR_ERR_CHECK_MATCHING_SIZES_HPP
+#define STAN_MATH_PRIM_ARR_ERR_CHECK_MATCHING_SIZES_HPP
 
 #include <stan/math/prim/scal/err/domain_error.hpp>
 #include <stan/math/prim/scal/err/check_size_match.hpp>
