+++ stan/math/prim/arr/fun/inverse_softmax.hpp
@@ -1,7 +1,7 @@
-#ifndef STAN_MATH_PRIM_SCAL_FUN_INVERSE_SOFTMAX_HPP
-#define STAN_MATH_PRIM_SCAL_FUN_INVERSE_SOFTMAX_HPP
+#ifndef STAN_MATH_ARR_SCAL_FUN_INVERSE_SOFTMAX_HPP
+#define STAN_MATH_ARR_SCAL_FUN_INVERSE_SOFTMAX_HPP
 
-#include <stan/math/prim/mat/err/check_matching_sizes.hpp>
+#include <stan/math/prim/arr/err/check_matching_sizes.hpp>
 #include <cmath>
 
 namespace stan {
